move base recipe from zouwu to automl
